requirements_detector.txt
requests
python-dotenv
PyPDF2
python-docx